Filter notes by authenticated user

+++ app/Http/Controllers/NoteController.php
@@ -13,7 +13,10 @@ class NoteController extends Controller
      */
     public function index()
     {
-        $notes = Note::query()->orderBy("created_at", "desc")->paginate(10);
+        $notes = Note::query()
+          ->where("user_id", request()->user()->id)
+          ->orderBy("created_at", "desc")
+          ->paginate(10);
         return view("note.index", ["notes" => $notes]);
     }
 
@@ -22,7 +25,7 @@ class NoteController extends Controller
      */
     public function create()
     {
-        return view("note.create");
+      return view("note.create");
     }
 
     /**
@@ -34,7 +37,7 @@ class NoteController extends Controller
           "note" => ["required", "string"]
         ]);
 
-        $data["user_id"] = 1;
+        $data["user_id"] = $request->user()->id;
         $note = Note::create($data);
 
         return to_route("note.show", $note)->with("message", "Note was created");
@@ -45,7 +48,10 @@ class NoteController extends Controller
      */
     public function show(Note $note)
     {
-        return view("note.show", ["note" => $note]);
+      if ($note->user_id !== request()->user()->id) {
+        abort(403);
+      }
+      return view("note.show", ["note" => $note]);
     }
 
     /**
@@ -53,7 +59,10 @@ class NoteController extends Controller
      */
     public function edit(Note $note)
     {
-        return view("note.edit", ["note" => $note]);
+      if ($note->user_id !== request()->user()->id) {
+        abort(403);
+      }
+      return view("note.edit", ["note" => $note]);
     }
 
     /**
@@ -61,13 +70,16 @@ class NoteController extends Controller
      */
     public function update(Request $request, Note $note)
     {
-        $data = $request->validate([
-          "note" => ["required", "string"]
-        ]);
+      if ($note->user_id !== request()->user()->id) {
+        abort(403);
+      }
+      $data = $request->validate([
+        "note" => ["required", "string"]
+      ]);
 
-        $note->update($data);
+      $note->update($data);
 
-        return to_route("note.show", $note)->with("message", "Note has updated");
+      return to_route("note.show", $note)->with("message", "Note has updated");
     }
 
     /**
@@ -75,8 +87,11 @@ class NoteController extends Controller
      */
     public function destroy(Note $note)
     {
-        $note->delete();
+      if ($note->user_id !== request()->user()->id) {
+        abort(403);
+      }
+      $note->delete();
 
-        return to_route("note.index")->with("message", "Note has deleted");
+      return to_route("note.index")->with("message", "Note has deleted");
     }
 }
